-) Add Abstract parent to dvas.data.data.MultiProfile -) Remove time index from DB -) Add abstract parent to dvas.data.strategy.load.LoadProfileStrategy -) Modify dvas.data.strategy.load.LoadStrategyAbstract._fetch method to more generic

+++ src/dvas/database/model.py
@@ -136,5 +136,5 @@ class Data(MetadataModel):
     event_info = ForeignKeyField(
         EventsInfo,
         backref='datas')
-    index = FloatField(null=False)
+    index = IntegerField(null=False)
     value = FloatField(null=True)
